--- src/main/java/Task.java
@@ -1,35 +1,0 @@
-public class Task {
-    protected String taskName;
-    protected boolean isDone;
-
-    public Task(String taskName) {
-        this.taskName = taskName;
-        this.isDone = false;
-    }
-
-    public Task(String taskName, boolean isDone) {
-        this.taskName = taskName;
-        this.isDone = isDone;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "X" : " ");
-    }
-
-    public void markAsDone() {
-        this.isDone = true;
-    }
-
-    public void markAsNotDone() {
-        this.isDone = false;
-    }
-
-    public String getTaskName() {
-        return this.taskName;
-    }
-
-    @Override
-    public String toString() {
-        return "[" + getStatusIcon() + "] " + getTaskName();
-    }
-}
