+++ src/main/java/duke/DukeException.java
@@ -1,3 +1,4 @@
+package duke;
 public class DukeException extends RuntimeException {
 
     public DukeException(String e) {
