+++ src/main/java/duke/Storage.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
