+++ src/main/java/duke/TaskList.java
@@ -1,0 +1,31 @@
+package duke;
+
+import java.util.ArrayList;
+
+public class TaskList {
+    private static ArrayList<Task> taskList;
+    private int taskListSize;
+
+    public TaskList() {
+        this.taskList = new ArrayList<>();
+        this.taskListSize = 0;
+    }
+
+    public static void add(Task task) {
+        taskList.add(task);
+    }
+
+    public int getTaskListSize() {
+        return taskList.size();
+    }
+
+    public Task get(int i) {
+        return taskList.get(i);
+    }
+
+
+    public void remove(int numToDelete) {
+        Task currTask = this.taskList.get(numToDelete - 1);
+        this.taskList.remove(numToDelete - 1);
+    }
+}
